Create verse parameter
This is intended to allow a targeted search of texts based on book,
chapter, and verse---or possibly just book and chapter, or even just
book, but that'll be a later enhancement. The default usage mode for
the program should be to provide a verse designator and have it print
the verse in one or more translations.