Search implemented
I'm using basically edit distance for equal-length substrings on a by-
verse basis, so don't bother trying to search for any string that is
split across verses. Also, it's entirely possible that, because the
Bible is long, it's entirely possible that your search string will
work but won't appear in the results because the phrase you searched
for will appear more than ten times before we get  to the part of the
Bible you were thinking of.