Smarter fuzzy search
The fuzzy search now only compares against windows beginning with
matching characters if those characters are found at the beginning of
words. For instance:

text:  how now brown cow
query: ow

The searcher will find ZERO matching windows in the search string,
because the search requires that the window start with the beginning
of a word.